fix: BankAccount, CurrentAccount, Transactable, SavingsAccount, main

+++ src/base/Transactable.java
@@ -1,6 +1,8 @@
 package base;
 
+import java.math.BigDecimal;
+
 public interface Transactable {
 
-    public boolean transfer(BankAccount target, int amount);
+    public boolean transfer(BankAccount target, BigDecimal amount);
 }
